+++ tests/unit-global/es.string.raw.js
@@ -8,6 +8,7 @@ QUnit.test('String.raw', assert => {
   assert.same(raw({ raw: ['Hi\\n', '!'] }, 'Bob'), 'Hi\\nBob!', 'raw is array');
   assert.same(raw({ raw: 'test' }, 0, 1, 2), 't0e1s2t', 'raw is string');
   assert.same(raw({ raw: 'test' }, 0), 't0est', 'lacks substituting');
+  assert.same(raw({ raw: [] }), '', 'empty template');
 
   if (typeof Symbol == 'function' && !Symbol.sham) {
     assert.throws(() => raw({ raw: [Symbol()] }, 0), TypeError, 'throws on symbol #1');