+++ tests/unit-pure/es.string.raw.js
@@ -9,6 +9,7 @@ QUnit.test('String.raw', assert => {
   assert.same(raw({ raw: ['Hi\\n', '!'] }, 'Bob'), 'Hi\\nBob!', 'raw is array');
   assert.same(raw({ raw: 'test' }, 0, 1, 2), 't0e1s2t', 'raw is string');
   assert.same(raw({ raw: 'test' }, 0), 't0est', 'lacks substituting');
+  assert.same(raw({ raw: [] }), '', 'empty template');
 
   /* eslint-disable es/no-symbol -- safe */
   if (typeof Symbol == 'function') {