+++ tests/unit-pure/esnext.string.cooked.js
@@ -8,6 +8,7 @@ QUnit.test('String.cooked', assert => {
   assert.same(cooked(['Hi\\n', '!'], 'Bob'), 'Hi\\nBob!', 'template is an array');
   assert.same(cooked('test', 0, 1, 2), 't0e1s2t', 'template is a string');
   assert.same(cooked('test', 0), 't0est', 'lacks substituting');
+  assert.same(cooked([]), '', 'empty template');
 
   if (typeof Symbol == 'function' && !Symbol.sham) {
     assert.throws(() => cooked([Symbol()]), TypeError, 'throws on symbol #1');